requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
idna==2.9
pytube3==9.6.4
requests==2.23.0
soupsieve==2.0
typing-extensions==3.7.4.1
urllib3==1.25.8
